Update Editor Select Entity
Update Editor Select Entity

+++ BunnyEngine/src/BunnyEngine/Renderer/Framebuffer.h
@@ -55,7 +55,7 @@ namespace BE {
 		virtual void UnBind() =0;
 
 		virtual void Resize(uint32_t width, uint32_t height) = 0;
-		virtual float ReadPixel(uint32_t attachmentIndex, int x, int y) = 0;
+		virtual int ReadPixel(uint32_t attachmentIndex, int x, int y) = 0;
 
 		virtual void ClearAttachment(uint32_t attachmentIndex, int value) = 0;
 		virtual void BindAttachment(uint32_t attachmentIndex, int value, bool Depth = 0) = 0;
